Fixes totalPizzas text string issue in CONTROLLER from changing after getting the total pizzas number

+++ app/src/main/java/com/zybooks/pizzaparty/PizzaCalculator.java
@@ -67,7 +67,7 @@ public class PizzaCalculator {
             pizzaSlices = 12;
         }
 
-        return (int) Math.ceil(mPartySize * slicesPerPerson / pizzaSlices);
+        return (int) Math.ceil(mPartySize * slicesPerPerson / (double) pizzaSlices);
 //      return (int) Math.ceil(mPartySize * slicesPerPerson / (double) SLICES_PER_PIZZA);
     }
 }
